+++ tests/testthat/test-binomial.R
@@ -33,7 +33,6 @@ test_that("regularized slope logistic regression picks out correct features", {
   k <- 3
 
   x <- matrix(rnorm(p * n), n, p)
-  # x <- scale(x)
   beta <- double(p)
   nz <- sample(p, k)
 
@@ -46,4 +45,21 @@ test_that("regularized slope logistic regression picks out correct features", {
   slope_fit <- SLOPE(x, y, family = "binomial", alpha = 1 / sqrt(n))
 
   expect_setequal(nz, which(slope_fit$nonzeros[[1]]))
+})
+
+test_that("logistic regression works on the glioma dataset", {
+  set.seed(825)
+
+  glioma_fit <- SLOPE(
+    x = glioma$x,
+    y = glioma$y,
+    family = "binomial",
+    alpha = 1e-2
+  )
+
+  y_hat <- predict(glioma_fit, glioma$x, type = "class")
+
+  accuracy <- mean(y_hat == glioma$y)
+
+  expect_equal(accuracy, 0.975, tolerance = 1e-1)
 })